feat(v2): add toc to blog posts
* refactor TOC into own component

* make layout consistent across blog pages
