+++ packages/docusaurus-theme-classic/src/theme/BlogPostPage/index.tsx
@@ -10,11 +10,13 @@ import React from 'react';
 import Layout from '@theme/Layout';
 import BlogPostItem from '@theme/BlogPostItem';
 import BlogPostPaginator from '@theme/BlogPostPaginator';
+import TOC from '@theme/TOC';
 
 function BlogPostPage(props): JSX.Element {
   const {content: BlogPostContents} = props;
   const {frontMatter, metadata} = BlogPostContents;
   const {title, description, nextItem, prevItem, editUrl} = metadata;
+  const {hide_table_of_contents: hideTableOfContents} = frontMatter;
 
   return (
     <Layout title={title} description={description}>
@@ -55,6 +57,11 @@ function BlogPostPage(props): JSX.Element {
                 </div>
               )}
             </div>
+            {!hideTableOfContents && BlogPostContents.rightToc && (
+              <div className="col col--2">
+                <TOC headings={BlogPostContents.rightToc} />
+              </div>
+            )}
           </div>
         </div>
       )}
