+++ packages/docusaurus-theme-classic/src/theme/DocItem/styles.module.css
@@ -29,37 +29,7 @@
   }
 }
 
-.tableOfContents {
-  display: inherit;
-  max-height: calc(100vh - (var(--ifm-navbar-height) + 2rem));
-  overflow-y: auto;
-  position: sticky;
-  top: calc(var(--ifm-navbar-height) + 2rem);
-}
-
-.tableOfContents::-webkit-scrollbar {
-  width: 7px;
-}
-
-.tableOfContents::-webkit-scrollbar-track {
-  background: #f1f1f1;
-  border-radius: 10px;
-}
-
-.tableOfContents::-webkit-scrollbar-thumb {
-  background: #888;
-  border-radius: 10px;
-}
-
-.tableOfContents::-webkit-scrollbar-thumb:hover {
-  background: #555;
-}
-
 @media only screen and (max-width: 996px) {
-  .tableOfContents {
-    display: none;
-  }
-
   .docItemContainer {
     padding: 0 0.3rem;
   }
